tools/data-generator/requirement.txt
Faker
faker-ecommerce-provider
psycopg2